refactor pages to countries
